+++ hardware.h
@@ -1,5 +1,5 @@
 /******************************************************************************\
- * Copyright (c) 2020-2024
+ * Copyright (c) 2020-2026
  * Author(s): Volker Fischer
  ******************************************************************************
  * This program is free software; you can redistribute it and/or modify it under
@@ -53,7 +53,7 @@ class Edrumulus_hardware
   void capture_samples(const int number_pads,
                        const int number_inputs[],
                        int       analog_pin[][MAX_NUM_PAD_INPUTS],
-                       int       sample_org[][MAX_NUM_PAD_INPUTS]);
+                       uint16_t  sample_org[][MAX_NUM_PAD_INPUTS]);
 
   void write_setting(const int pad_index, const int address, const byte value);
   byte read_setting(const int pad_index, const int address);
@@ -107,7 +107,7 @@ class Edrumulus_hardware
   void capture_samples(const int number_pads,
                        const int number_inputs[],
                        int       analog_pin[][MAX_NUM_PAD_INPUTS],
-                       int       sample_org[][MAX_NUM_PAD_INPUTS]);
+                       uint16_t  sample_org[][MAX_NUM_PAD_INPUTS]);
 
   void write_setting(const int, const int, const byte) {}; // not supported
   byte read_setting(const int, const int) { return 0; };   // not supported
